Working in Loading response to medals page Need to have ready a mockup to start creating a skeleton
