+++ src/app/(main)/medals/loading.tsx
@@ -1,0 +1,9 @@
+"use client"
+
+
+export default function Loading() {
+
+
+return "Loading..."
+
+}
