+++ src/app/api/medals.ts
@@ -1,11 +1,12 @@
 import { Medal } from "../../types/medals"
 
 
-
 export async function getMedals() : Promise<Medal[]> {
 
+  const BASE_URL = 'http://localhost:3004'
 
-  return await fetch('http://localhost:3004/api/medals')
+
+  return await fetch(`${BASE_URL}/api/medals`)
   .then(res => res.json())
